chore: devide column search from bodywrapper
