+++ src/components/BodyWrapper.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect, useRef, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import {
   Box,
   Flex,
@@ -9,11 +9,6 @@ import {
   Th,
   Thead,
   Tr,
-  Text,
-  InputGroup,
-  InputLeftElement,
-  Icon,
-  Input,
 } from '@chakra-ui/react';
 import { DataInterface } from '../const/types';
 import {
@@ -25,7 +20,6 @@ import {
   HeaderGroup,
   SortingState,
 } from '@tanstack/react-table';
-
 import {
   FilterContext,
   TableStatusContext,
@@ -33,8 +27,7 @@ import {
 import { UpDownIcon, ChevronDownIcon, ChevronUpIcon } from '@chakra-ui/icons';
 
 import { getFilteredData } from '../Data/Api';
-import { FiSearch } from 'react-icons/fi';
-import { t } from 'i18next';
+import ColumnSearch from './bodycomponents/ColumnSearch';
 
 const BodyWrapper = ({ columns }: { columns: any }) => {
   const { filterTerm, setFilterTerm } = useContext(FilterContext);
@@ -50,39 +43,7 @@ const BodyWrapper = ({ columns }: { columns: any }) => {
   const columnResizeMode: ColumnResizeMode = 'onChange';
   const [sorting, setSorting] = useState<SortingState>([]);
 
-  const placeholder = t('Search') || 'Search';
-
   const [data, setData] = useState<DataInterface[]>([]);
-
-  const [inputValues, setInputValues] = useState<Record<string, string>>({});
-
-  const handleSearchKeyDown = (
-    event: React.KeyboardEvent<HTMLInputElement>,
-    id: string | undefined
-  ) => {
-    if (event.key === 'Enter') {
-      setFilterTerm({ ...filterTerm, individualSearchTerm: inputValues });
-    }
-  };
-
-  const handleSearchOnChange = (
-    event: React.ChangeEvent<HTMLInputElement>,
-    id: string | undefined
-  ) => {
-    if (id) {
-      const value = (event.target as HTMLInputElement).value;
-      if (!value) {
-        const temp = inputValues;
-        delete temp[id];
-        setInputValues(temp);
-      } else {
-        setInputValues({
-          ...inputValues,
-          [id]: value,
-        });
-      }
-    }
-  };
 
   useEffect(() => {
     const fields: string[] = [];
@@ -229,44 +190,10 @@ const BodyWrapper = ({ columns }: { columns: any }) => {
                           opacity="0"
                         />
                       </Box>
-                      {header.column.columnDef.header !== '' && (
-                        <InputGroup maxW="sm">
-                          <InputLeftElement>
-                            {!isLoading ? (
-                              <Icon
-                                as={FiSearch}
-                                color="fg.muted"
-                                boxSize="5"
-                                cursor="pointer"
-                                onClick={() =>
-                                  setFilterTerm({
-                                    ...filterTerm,
-                                    individualSearchTerm: inputValues,
-                                  })
-                                }
-                              />
-                            ) : (
-                              <Spinner size="sm" />
-                            )}
-                          </InputLeftElement>
-                          <Input
-                            placeholder={placeholder}
-                            onKeyDown={(event) => {
-                              handleSearchKeyDown(
-                                event,
-                                header.column.columnDef.id
-                              );
-                            }}
-                            disabled={isLoading}
-                            onChange={(event) => {
-                              handleSearchOnChange(
-                                event,
-                                header.column.columnDef.id
-                              );
-                            }}
-                          />
-                        </InputGroup>
-                      )}
+                      {header.column.columnDef.header !== '' &&
+                        header.column.columnDef.id && (
+                          <ColumnSearch id={header.column.columnDef.id} />
+                        )}
                     </Flex>
                   </Th>
                 ))}
